refactor: remove the DB Manager page in favour of the DB Viewer
Drops the DatabaseWidget stack page from music_manager.ui (with its
"DB Manager" menu action and custom-widget declaration) and deletes
ui/db_window_widget.py. The shared table-view helpers the DB Viewer
relied on (CenterAlignDelegate, setup_data_view, on_row_clicked) move
into db_media_window.py. The DB Viewer menu action now targets stack
index 1. README updated to match.

The legacy DatabaseWindow popup (ui/db_window.py, Ctrl+D) is untouched.
